Remove Request.user declarations (#486)
I actually moved them to the test file that uses this property to keep the typescript compiler happy.
I added the tests & docs folder to the .npmignore to avoid conflict and because they shouldn't be published (as far as I know).
No template modification was needed because `request` has an any type there.